beautification: move wang landau function definition to helper file
